os/ZFS: move to os/fs
