+++ src/os/Makefile.am
@@ -99,10 +99,10 @@ noinst_HEADERS += \
 	os/ObjectStore.h
 
 if WITH_LIBZFS
-libos_zfs_a_SOURCES = os/ZFS.cc
+libos_zfs_a_SOURCES = os/fs/ZFS.cc
 libos_zfs_a_CXXFLAGS = ${AM_CXXFLAGS} ${LIBZFS_CFLAGS}
 noinst_LIBRARIES += libos_zfs.a
-noinst_HEADERS += os/ZFS.h
+noinst_HEADERS += os/fs/ZFS.h
 endif
 
 ceph_bluefs_tool_SOURCES = os/bluestore/bluefs_tool.cc
